requirements.txt
flask
bcrypt
celery
redis
spotipy==2.25.1
python-dotenv
filetype
gunicorn
